test: split integration and unit tests
Use Gradle's test suites to separate the unit and integration tests.
This feature is incubating so may need some tweaks between Gradle
versions, but the functionality is permanent rather than experimental.

Integration tests will be ran with `gradlew integrationTest`.

The Gradle dependencies should be updated to move test dependencies in
to the test suite configuration instead of the main dependencies block.

TIS21-5582
TIS21-5664