refactor: action model `due` field
The Action model/DTO has a single `due` field, which should be split in
to `availableFrom` to represent when an action is completable and
`dueBy` to indicate the "deadline" for the action completion.

Update the Action mapper to set the new date fields to appropriate
values. With Placement dueBy matching the start date and availableFrom
being based on the 12 week window. Programme should set avilableFrom to
the current day to be instantly available and the dueBy to the Programme
start date.

Update the test classes to cover the refactoring and date field changes.

TIS21-5582