chore: disable Sentry on preprod
The preprod service is creating a large amount of Sentry errors due to
duplicate indexes.
Until there is an answer/solution to the problem Sentry reporting should
be disabled to avoid using up all of our allowance.

Create properties for a `preprod` profile which disables sentry.
Update the task definition to set the `SPRING_PROFILES_ACTIVE` env var
based on the environment.

NO-TICKET